clean up in structure
