+++ Vostok.Singular.Core.Tests/IclResolverTests.cs
@@ -1,9 +1,9 @@
 using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
-using Vostok.Singular.Core.Idempotency.Identifier;
-using Vostok.Singular.Core.Identifier;
-using Vostok.Singular.Core.Settings;
+using Vostok.Singular.Core.Idempotency;
+using Vostok.Singular.Core.Idempotency.Icl;
+using Vostok.Singular.Core.Idempotency.Icl.Settings;
 
 namespace Vostok.Singular.Core.Tests
 {
