+++ Vostok.Singular.Core.Tests/IclRuleMatcherTests.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using FluentAssertions;
 using NUnit.Framework;
-using Vostok.Singular.Core.Idempotency.Identifier;
-using Vostok.Singular.Core.Identifier;
+using Vostok.Singular.Core.Idempotency;
+using Vostok.Singular.Core.Idempotency.Icl;
 
 namespace Vostok.Singular.Core.Tests
 {
