+++ Vostok.Singular.Core/Idempotency/BlackList/Settings/IIdempotencySignsProvider.cs
@@ -1,6 +1,4 @@
-﻿using Vostok.Singular.Core.Idempotency.Settings;
-
-namespace Vostok.Singular.Core.Idempotency
+﻿namespace Vostok.Singular.Core.Idempotency.BlackList.Settings
 {
     internal interface IIdempotencySignsProvider
     {
