+++ Vostok.Singular.Core/Idempotency/BlackList/Settings/IdempotencySignsProvider.cs
@@ -1,19 +1,19 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Vostok.ClusterConfig.Client;
 using Vostok.Configuration;
 using Vostok.Configuration.Abstractions;
 using Vostok.Configuration.Sources.ClusterConfig;
 using Vostok.Configuration.Sources.Json;
-using Vostok.Singular.Core.Settings;
 
-namespace Vostok.Singular.Core
+namespace Vostok.Singular.Core.Idempotency.BlackList.Settings
 {
-    internal class IclSettingsProvider : IIclSettingsProvider
+    internal class IdempotencySignsProvider : IIdempotencySignsProvider
     {
         private readonly IConfigurationSource source;
         private readonly string servicePath;
+        private static readonly NonIdempotencySignsSettings EmptySigns = new NonIdempotencySignsSettings {Signs = new List<NonIdempotencySignSettings>(0)};
 
-        public IclSettingsProvider(string serviceName, string configurationPathPrefix = SingularClientConstants.ServicesConfigurationNamePrefix)
+        public IdempotencySignsProvider(string serviceName, string configurationPathPrefix = SingularClientConstants.ServicesConfigurationNamePrefix)
         {
             servicePath = $"{configurationPathPrefix}{serviceName}.json";
             source = new ClusterConfigSource(
@@ -23,12 +23,12 @@ namespace Vostok.Singular.Core
                 });
         }
 
-        public List<IdempotencyRuleSetting> Get()
+        public NonIdempotencySignsSettings Get()
         {
             if (ClusterConfigClient.Default.Get(servicePath) == null)
-                return new List<IdempotencyRuleSetting>(0);
+                return EmptySigns;
 
-            return ConfigurationProvider.Default.Get<IdempotencyControlListSetting>(source).Rules;
+            return ConfigurationProvider.Default.Get<NonIdempotencyServiceSettings>(source).NonIdempotencySigns;
         }
     }
 }
