+++ Vostok.Singular.Core/Idempotency/BlackList/Settings/NonIdempotencySignSettings.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Vostok.Singular.Core.Idempotency.Settings
+namespace Vostok.Singular.Core.Idempotency.BlackList.Settings
 {
     [Serializable]
     internal class NonIdempotencySignSettings
