+++ Vostok.Singular.Core/Idempotency/Icl/IdempotencyControlRule.cs
@@ -1,7 +1,6 @@
-using Vostok.Singular.Core.Idempotency.Identifier;
-using Vostok.Singular.Core.Settings;
+using Vostok.Singular.Core.Idempotency.Icl.Settings;
 
-namespace Vostok.Singular.Core.Identifier
+namespace Vostok.Singular.Core.Idempotency.Icl
 {
     internal class IdempotencyControlRule
     {
