+++ Vostok.Singular.Core/Idempotency/Icl/Settings/IIclCache.cs
@@ -1,6 +1,4 @@
-using Vostok.Singular.Core.Identifier;
-
-namespace Vostok.Singular.Core
+namespace Vostok.Singular.Core.Idempotency.Icl.Settings
 {
     internal interface IIclCache
     {
