+++ Vostok.Singular.Core/Idempotency/Icl/Settings/IclCache.cs
@@ -1,11 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using Vostok.Commons.Collections;
-using Vostok.Singular.Core.Idempotency.Identifier;
-using Vostok.Singular.Core.Identifier;
-using Vostok.Singular.Core.Settings;
 
-namespace Vostok.Singular.Core
+namespace Vostok.Singular.Core.Idempotency.Icl.Settings
 {
     internal class IclCache : IIclCache
     {
